Add files via upload
About Dataset
This is a fitness tracker product dataset for the Indian market that includes various products from various brands, along with their specifications, ratings, and reviews. Web scraping was used to collect data from e-commerce websites such as Flipkart and Amazon.

Inspiration
This dataset could be used to find answers to some interesting questions like -

Is there a significant demand for fitness trackers in the Indian market?
Information on the top 5 brands for fitness bands and smartwatches
Is there a correlation between the prices and product specifications, ratings, etc?
Different types of fitness trackers, and their price segments for different users
This dataset contains 451 samples with 16 attributes. There are some missing values in this dataset.